Move Config initialization & error handling to scaffolding.
