+++ Cargo.toml
@@ -8,7 +8,6 @@ version = "0.1.0"
 base64 = "0.1.1"
 mime = "0.1.3"
 unicode-width = "0.1.3"
-toml = "0.1"
 
 [dependencies.notty-encoding]
 git = "https://github.com/withoutboats/notty-encoding"
